+++ Statistics/Banking/PrivatBank.cs
@@ -17,7 +17,7 @@ using Visualization.Banking;
 
 namespace Statistics.Banking
 {
-	public class PrivatBank : BindableBase, IFundsStorage, IRecordsProvider, IStorage<decimal>
+	public class PrivatBank : BindableBase, IFundsStorage, IRecordsProvider, IStorage<int>
 	{
 		private readonly string historyUrl = "https://api.privatbank.ua/p24api/rest_fiz";
 		private readonly string balanceUrl = "https://api.privatbank.ua/p24api/balance";
@@ -87,7 +87,7 @@ namespace Statistics.Banking
 				</data>
 			</request>";
 
-		public decimal Value { get; set; }
+		public int Value { get; set; }
 
 		public PrivatBank()
 		{
@@ -100,7 +100,7 @@ namespace Statistics.Banking
 		{
 			var file = PrepareData();
 			var responce = await SendData(balanceUrl, file);
-			Value = Parser.ParseBalance(responce);
+			Value = (int) Parser.ParseBalance(responce);
 
 			OnPropertyChanged("Value");
 		}
